requirement.txt
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
oauth2client==4.1.3
selenium==3.141.0
pytest-xdist==1.29.0
pytest==5.2.0
